Move ftp and http staging into own files and out of DM
The intention here is to move the particular staging
mechanism out of the DM and into something configurable
along the lines of GlobusScheme
