+++ parsl/data_provider/ftp.py
@@ -1,0 +1,23 @@
+import ftplib
+import os
+
+from parsl import python_app
+
+
+def _ftp_stage_in(working_dir, outputs=[]):
+    file = outputs[0]
+    if working_dir:
+        os.makedirs(working_dir, exist_ok=True)
+        file.local_path = os.path.join(working_dir, file.filename)
+    else:
+        file.local_path = file.filename
+    with open(file.local_path, 'wb') as f:
+        ftp = ftplib.FTP(file.netloc)
+        ftp.login()
+        ftp.cwd(os.path.dirname(file.path))
+        ftp.retrbinary('RETR {}'.format(file.filename), f.write)
+        ftp.quit()
+
+
+def _ftp_stage_in_app(dm, executor):
+    return python_app(executors=[executor], data_flow_kernel=dm.dfk)(_ftp_stage_in)
