+++ parsl/data_provider/http.py
@@ -1,0 +1,22 @@
+import os
+import requests
+
+from parsl import python_app
+
+
+def _http_stage_in(working_dir, outputs=[]):
+    file = outputs[0]
+    if working_dir:
+        os.makedirs(working_dir, exist_ok=True)
+        file.local_path = os.path.join(working_dir, file.filename)
+    else:
+        file.local_path = file.filename
+    resp = requests.get(file.url, stream=True)
+    with open(file.local_path, 'wb') as f:
+        for chunk in resp.iter_content(chunk_size=1024):
+            if chunk:
+                f.write(chunk)
+
+
+def _http_stage_in_app(dm, executor):
+    return python_app(executors=[executor], data_flow_kernel=dm.dfk)(_http_stage_in)
